\name{clevRvisShiny}
\alias{clevRvisShiny}
\title{
A user interface to perform all analyses with clevRvis.
}
\description{
clevRvis is a tool that allows you to visualize changes in the subclonal
architecture of tumors. Simple tree visualization (shark plot) a more detailed
visualization (dolphin plot) and an allele-aware visualization (plaice plot)
are available.
Moreover, this tool provides fully automatic algorithms for time point
interpolation and therapy effect estimation in the presence of lacking input
data.
}
\usage{
clevRvisShiny()
}
\arguments{
None
}
\details{
Detailed information on the usage of clevRvisShiny are available within the
shiny app (see Tutorial).
}
\value{
None
}
\author{
Sarah Sandmann <sarah.sandmann@uni-muenster.de>
}


\examples{
\donttest{
clevRvisShiny()
}}
